Change scratch.php to class base code, but doesn't use class field.
